Add warning about including display-essential js in async loader.
Since the async loader much more strongly defers the loading of the
javascript, the page can display much faster. However, if any of
the javascirpt is required to properly display the page, then the
method will result in a much longer flash of unstyled content.
This warning alerts the user of that fact.